SSH fingerprints! Also a realloc fix.
The idea is, a SSH server sends the fingerprint of your client public
key in a line by itself. Then the normal protocol resumes afterwards.
What it means for λμ is that if the fingerprint on the tqueue has not
been set, assume the first line is a fingerprint. Makes the login screen
not respond for one line if you don't use a special ssh to get to this
moo via port forwarding!

Uh, and it's all disabled unless you specify --with-ssh-fingerprints in
configure so I can use it, and nobody else needs to. Except the realloc
thing, that's just common sense.